Update Main_file
So turns out the way I was calculating the plant was actually just wrong. This way, the program is basically "plotting" the roots on a root locus graph and working out the response from there. This method assumes the transfer function is coprime. It works but the pid control hasn't been tested so that could just not work now. 

There is a very slight issue with the fact the reponse suddenly changes on the first point but other than that, it works. :)

No reverse laplace required.

This code has no protection in regards to there being more zero's than poles (which can't happen in real life) so don't do that. 
(poles are roots on the denominator and zero's are roots on the numerator)