Update Main_file
Starting the process to find a way to determine arbitrary constants with boundary conditions. 

because the components of the response can only be exponential or trig or both, you can limit what needs to be calculated and sidestep actually having to calculate derivatives. 

this forms a set of linear equations which will then be solved by inverting a matrix and these coefficients will be applied to follow boundary conditions. Hopefully