Update siso_simulation_functions
Added the ability to plot root locus's of the plant. 
Once again, you aren't supposed to have more zeros than poles so try and avoid that. There are only 3 different types of lines so plot more at the risk of it looking messy. 

It goes from 0  to 1000 for values of k in steps of 0.1
If you want more, change in in this library but at the risk of it obviously taking longer